Refactoring jsx to tsx, externalizing the auth config, adding types, upgrading libraries
