+++ .gitignore
@@ -1,3 +1,5 @@
+config.json
+
 # Logs
 logs
 *.log
