+++ package.json
@@ -3,17 +3,7 @@
   "version": "0.1.0",
   "description": "A React single-page application calling Microsoft Graph API using MSAL.js (w/ AAD v2 endpoint)",
   "private": true,
-  "dependencies": {
-    "@azure/msal-browser": "^2.15.0",
-    "@azure/msal-react": "^1.0.1",
-    "bootstrap": "^4.5.3",
-    "jquery": "^3.5.1",
-    "popper.js": "^1.16.1",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^17.0.0",
-    "react-scripts": "^4.0.2"
-  },
+  "dependencies": {},
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build"
@@ -29,5 +19,15 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "typescript": "^4.4.4",
+    "@azure/msal-browser": "^2.18.0",
+    "@azure/msal-react": "^1.1.0",
+    "bootstrap": "^4.5.3",
+    "react": "^17.0.1",
+    "react-bootstrap": "^1.3.0",
+    "react-dom": "^17.0.2",
+    "react-scripts": "^4.0.3"
   }
 }
