+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { AuthenticatedTemplate, UnauthenticatedTemplate, useMsal } from "@azure/msal-react";
-import { loginRequest } from "./authConfig";
+import { GraphData, loginRequest } from "./authConfig";
 import { PageLayout } from "./components/PageLayout";
 import { ProfileData } from "./components/ProfileData";
 import { callMsGraph } from "./graph";
@@ -10,9 +10,9 @@ import "./styles/App.css";
 /**
  * Renders information about the signed-in user or a button to retrieve data about the user
  */
-const ProfileContent = () => {
+const ProfileContent: React.FC = () => {
     const { instance, accounts } = useMsal();
-    const [graphData, setGraphData] = useState(null);
+    const [graphData, setGraphData] = useState<GraphData>();
 
     function RequestProfileData() {
         // Silently acquires an access token which is then attached to a request for MS Graph data
@@ -27,7 +27,7 @@ const ProfileContent = () => {
     return (
         <>
             <h5 className="card-title">Welcome {accounts[0].name}</h5>
-            {graphData ? 
+            {graphData ?
                 <ProfileData graphData={graphData} />
                 :
                 <Button variant="secondary" onClick={RequestProfileData}>Request Profile Information</Button>
@@ -39,7 +39,7 @@ const ProfileContent = () => {
 /**
  * If a user is authenticated the ProfileContent component above is rendered. Otherwise a message indicating a user is not authenticated is rendered.
  */
-const MainContent = () => {    
+const MainContent: React.FC = () => {
     return (
         <div className="App">
             <AuthenticatedTemplate>
