+++ src/components/PageLayout.tsx
@@ -12,9 +12,9 @@ import { SignOutButton } from "./SignOutButton";
 
 /**
  * Renders the navbar component with a sign-in or sign-out button depending on whether or not a user is authenticated
- * @param props 
+ * @param props
  */
-export const PageLayout = (props) => {
+export const PageLayout: React.FC = (props) => {
     const isAuthenticated = useIsAuthenticated();
 
     return (
@@ -23,7 +23,7 @@ export const PageLayout = (props) => {
                 <a className="navbar-brand" href="/">Microsoft Identity Platform</a>
                 { isAuthenticated ? <SignOutButton /> : <SignInButton /> }
             </Navbar>
-            <h5><center>Welcome to the Microsoft Authentication Library For Javascript - React Quickstart</center></h5>
+            <h5 className="text-center">Welcome to the Microsoft Authentication Library For Javascript - React Quickstart</h5>
             <br />
             <br />
             {props.children}
