+++ src/config.example.json
@@ -1,0 +1,9 @@
+{
+  "clientId": "Enter_the_Application_Id_Here",
+  "authority": "Enter_the_Cloud_Instance_Id_HereEnter_the_Tenant_Info_Here",
+  "redirectUri": "Enter_the_Redirect_Uri_Here",
+  "scopes": [
+    "User.Read"
+  ],
+  "graphMeEndpoint": "Enter_the_Graph_Endpoint_Herev1.0/me"
+}
